requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
astor==0.8.1
asttokens==2.3.0
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.5
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
ctransformers==0.2.20
cycler==0.11.0
dataclasses-json==0.5.14
decorator==4.4.2
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.99.1
ffmpeg-python==0.2.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.252
langsmith==0.0.19
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.4
numpy==1.24.4
onnxruntime==1.15.1
openai==0.27.8
openai-whisper==20230314
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandasai==1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
posthog==3.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.23.4
pulsar-client==3.2.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.3.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.23.2
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
